+++ ios/ios/ViewModels/PerformRecipeViewModel.swift
@@ -16,62 +16,62 @@ class PerformRecipeViewModel {
   var recipe: Mealplanning_Recipe?
   var isLoading = false
   var errorMessage: String?
-  
+
   // Track which steps are completed (by step index)
   var completedSteps: Set<Int> = []
-  
+
   // Map from product ID to the step index that produces it
   var productIDToStepIndex: [String: Int] = [:]
-  
+
   private let recipeID: String
   private let authManager: AuthenticationManager
-  
+
   init(recipeID: String, authManager: AuthenticationManager) {
     self.recipeID = recipeID
     self.authManager = authManager
   }
-  
+
   func loadRecipe() async {
     isLoading = true
     errorMessage = nil
-    
+
     do {
       guard let clientManager = try? authManager.getClientManager() else {
         throw NSError(
           domain: "PerformRecipeViewModel", code: 1,
           userInfo: [NSLocalizedDescriptionKey: "Failed to get client manager"])
       }
-      
+
       // Get OAuth2 token (will refresh if needed)
       guard let oauth2Token = await authManager.getOAuth2AccessToken() else {
         throw NSError(
           domain: "PerformRecipeViewModel", code: 2,
           userInfo: [NSLocalizedDescriptionKey: "Failed to get OAuth2 access token"])
       }
-      
+
       let metadata = clientManager.authenticatedMetadata(accessToken: oauth2Token)
-      
+
       // Create request
       var request = Mealplanning_GetRecipeRequest()
       request.recipeID = recipeID
-      
+
       // Execute request
       let response = try await clientManager.client.mealPlanning.getRecipe(
         request,
         metadata: metadata,
         options: clientManager.defaultCallOptions
       )
-      
+
       self.recipe = response.result
       buildProductIDToStepIndexMapping()
     } catch {
       errorMessage = "Failed to load recipe: \(error.localizedDescription)"
       print("❌ Error loading recipe: \(error)")
     }
-    
+
     isLoading = false
   }
-  
+
   // Build a mapping from recipe step product IDs to the step index that produces them
   private func buildProductIDToStepIndexMapping() {
     guard let recipe = recipe else { return }
@@ -81,15 +81,15 @@ class PerformRecipeViewModel {
       }
     }
   }
-  
+
   // Check if a step can be checked off (all prerequisites are completed)
   func canCheckStep(_ stepIndex: Int) -> Bool {
     guard let recipe = recipe, stepIndex < recipe.steps.count else {
       return false
     }
-    
+
     let step = recipe.steps[stepIndex]
-    
+
     // Check all ingredients
     for ingredient in step.ingredients {
       if ingredient.hasRecipeStepProductID {
@@ -101,7 +101,7 @@ class PerformRecipeViewModel {
         }
       }
     }
-    
+
     // Check all instruments
     for instrument in step.instruments {
       if instrument.hasRecipeStepProductID {
@@ -113,7 +113,7 @@ class PerformRecipeViewModel {
         }
       }
     }
-    
+
     // Check all vessels
     for vessel in step.vessels {
       if vessel.hasRecipeStepProductID {
@@ -125,16 +125,16 @@ class PerformRecipeViewModel {
         }
       }
     }
-    
+
     return true
   }
-  
+
   // Toggle step completion
   func toggleStep(_ stepIndex: Int) {
     guard canCheckStep(stepIndex) else {
       return
     }
-    
+
     if completedSteps.contains(stepIndex) {
       // When unchecking, also uncheck all dependent steps
       uncheckStepAndDependents(stepIndex)
@@ -142,12 +142,12 @@ class PerformRecipeViewModel {
       completedSteps.insert(stepIndex)
     }
   }
-  
+
   // Recursively uncheck a step and all steps that depend on it
   private func uncheckStepAndDependents(_ stepIndex: Int) {
     guard let recipe = recipe else { return }
     completedSteps.remove(stepIndex)
-    
+
     // Find all steps that depend on this step's products
     let step = recipe.steps[stepIndex]
     for product in step.products {
@@ -156,9 +156,9 @@ class PerformRecipeViewModel {
         if dependentStepIndex <= stepIndex {
           continue  // Only check later steps
         }
-        
+
         var dependsOnProduct = false
-        
+
         // Check if any ingredient uses this product
         for ingredient in dependentStep.ingredients {
           if ingredient.hasRecipeStepProductID && ingredient.recipeStepProductID == product.id {
@@ -166,7 +166,7 @@ class PerformRecipeViewModel {
             break
           }
         }
-        
+
         // Check if any instrument uses this product
         if !dependsOnProduct {
           for instrument in dependentStep.instruments {
@@ -176,7 +176,7 @@ class PerformRecipeViewModel {
             }
           }
         }
-        
+
         // Check if any vessel uses this product
         if !dependsOnProduct {
           for vessel in dependentStep.vessels {
@@ -186,14 +186,14 @@ class PerformRecipeViewModel {
             }
           }
         }
-        
+
         if dependsOnProduct && completedSteps.contains(dependentStepIndex) {
           uncheckStepAndDependents(dependentStepIndex)
         }
       }
     }
   }
-  
+
   // Get the step at the given index
   func getStep(_ stepIndex: Int) -> Mealplanning_RecipeStep? {
     guard let recipe = recipe, stepIndex < recipe.steps.count else {
@@ -201,21 +201,21 @@ class PerformRecipeViewModel {
     }
     return recipe.steps[stepIndex]
   }
-  
+
   // Check if a step is completed
   func isStepCompleted(_ stepIndex: Int) -> Bool {
     return completedSteps.contains(stepIndex)
   }
-  
+
   // Get all prerequisite step indices for a given step
   func getPrerequisiteStepIndices(_ stepIndex: Int) -> [Int] {
     guard let recipe = recipe, stepIndex < recipe.steps.count else {
       return []
     }
-    
+
     var prerequisites: Set<Int> = []
     let step = recipe.steps[stepIndex]
-    
+
     // Check all ingredients
     for ingredient in step.ingredients {
       if ingredient.hasRecipeStepProductID {
@@ -225,7 +225,7 @@ class PerformRecipeViewModel {
         }
       }
     }
-    
+
     // Check all instruments
     for instrument in step.instruments {
       if instrument.hasRecipeStepProductID {
@@ -235,7 +235,7 @@ class PerformRecipeViewModel {
         }
       }
     }
-    
+
     // Check all vessels
     for vessel in step.vessels {
       if vessel.hasRecipeStepProductID {
@@ -245,13 +245,12 @@ class PerformRecipeViewModel {
         }
       }
     }
-    
+
     return Array(prerequisites).sorted()
   }
-  
+
   // Get the step index that produces a given product ID
   func getStepIndexForProductID(_ productID: String) -> Int? {
     return productIDToStepIndex[productID]
   }
 }
-
