+++ ios/ios/Views/Home/HomeView.swift
@@ -131,7 +131,7 @@ struct HomeView: View {
         .foregroundColor(.white)
         .cornerRadius(10)
       }
-      
+
       NavigationLink(destination: RecipeListView()) {
         HStack {
           Image(systemName: "book.closed")
@@ -363,12 +363,20 @@ struct PendingVoteCardContent: View {
             .foregroundColor(.secondary)
         }
         Spacer()
-        if hasVoted {
-          Image(systemName: "checkmark.circle.fill")
-            .foregroundColor(.green)
-        } else {
-          Image(systemName: "exclamationmark.circle.fill")
-            .foregroundColor(.orange)
+        HStack(spacing: 6) {
+          if hasVoted {
+            Image(systemName: "checkmark.circle.fill")
+              .foregroundColor(.green)
+            Text("Ballot submitted")
+              .font(.caption)
+              .foregroundColor(.green)
+          } else {
+            Image(systemName: "exclamationmark.circle.fill")
+              .foregroundColor(.orange)
+            Text("Vote requested")
+              .font(.caption)
+              .foregroundColor(.orange)
+          }
         }
       }
